Moved payload functions to new file
