+++ pkg/services/event/notification.go
@@ -3,12 +3,9 @@ package event
 import (
 	"go.uber.org/zap"
 
-	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
 	"github.com/gofrs/uuid"
 
-	"github.com/transcom/mymove/pkg/etag"
-	"github.com/transcom/mymove/pkg/gen/primemessages"
 	"github.com/transcom/mymove/pkg/models"
 	"github.com/transcom/mymove/pkg/services"
 	movetaskorder "github.com/transcom/mymove/pkg/services/move_task_order"
@@ -29,81 +26,6 @@ func isSourcePrime(event *Event) bool {
 
 // notificationSave saves the record in the webhook_notification table
 // If it fails, it will return an error.
-func notificationSave(event *Event, objectType string, payload *[]byte) error {
-	payloadString := string(*payload)
-	newNotification := models.WebhookNotification{
-		EventKey: string(event.EventKey),
-		Payload:  swag.String(payloadString),
-		Status:   models.WebhookNotificationPending,
-	}
-
-// PaymentRequestModelToPayload This is an example to log the data
-// Will need to move to payloads to model file.
-func PaymentRequestModelToPayload(paymentRequest *models.PaymentRequest) *primemessages.PaymentRequest {
-	if paymentRequest == nil {
-		return nil
-	}
-
-	paymentServiceItems := PaymentServiceItemsModelToPayload(&paymentRequest.PaymentServiceItems)
-	return &primemessages.PaymentRequest{
-		ID:                   strfmt.UUID(paymentRequest.ID.String()),
-		IsFinal:              &paymentRequest.IsFinal,
-		MoveTaskOrderID:      strfmt.UUID(paymentRequest.MoveTaskOrderID.String()),
-		PaymentRequestNumber: paymentRequest.PaymentRequestNumber,
-		RejectionReason:      paymentRequest.RejectionReason,
-		Status:               primemessages.PaymentRequestStatus(paymentRequest.Status),
-		PaymentServiceItems:  *paymentServiceItems,
-		ETag:                 etag.GenerateEtag(paymentRequest.UpdatedAt),
-	}
-}
-
-// PaymentServiceItemModelToPayload payload
-func PaymentServiceItemModelToPayload(paymentServiceItem *models.PaymentServiceItem) *primemessages.PaymentServiceItem {
-	if paymentServiceItem == nil {
-		return nil
-	}
-
-	payload := &primemessages.PaymentServiceItem{
-		ID:               strfmt.UUID(paymentServiceItem.ID.String()),
-		PaymentRequestID: strfmt.UUID(paymentServiceItem.PaymentRequestID.String()),
-		MtoServiceItemID: strfmt.UUID(paymentServiceItem.MTOServiceItemID.String()),
-		Status:           primemessages.PaymentServiceItemStatus(paymentServiceItem.Status),
-		RejectionReason:  paymentServiceItem.RejectionReason,
-		ETag:             etag.GenerateEtag(paymentServiceItem.UpdatedAt),
-	}
-
-	if paymentServiceItem.PriceCents != nil {
-		payload.PriceCents = swag.Int64(int64(*paymentServiceItem.PriceCents))
-	}
-
-	return payload
-}
-
-// PaymentServiceItemsModelToPayload payload
-func PaymentServiceItemsModelToPayload(paymentServiceItems *models.PaymentServiceItems) *primemessages.PaymentServiceItems {
-	if paymentServiceItems == nil {
-		return nil
-	}
-
-	payload := make(primemessages.PaymentServiceItems, len(*paymentServiceItems))
-
-	for i, p := range *paymentServiceItems {
-		payload[i] = PaymentServiceItemModelToPayload(&p)
-	}
-	return &payload
-}
-
-func isSourcePrime(event *Event) bool {
-	// We can assume the source is prime if the endpoint that caused the event was
-	// a prime endpoint. Support is considered non-prime endpoint.
-	// Generally we should err on the side of more notifications rather than fewer
-	// since we would not like them to miss an update.
-	apiName := GetEndpointAPI(event.EndpointKey)
-	if apiName == PrimeAPIName {
-		return true
-	}
-	return false
-}
 func notificationSave(event *Event, objectType string, payload *[]byte) error {
 	payloadString := string(*payload)
 	newNotification := models.WebhookNotification{
@@ -131,6 +53,10 @@ func notificationSave(event *Event, objectType string, payload *[]byte) error {
 	return nil
 
 }
+
+// checkAvailabilityToPrime returns true if the MTO is
+// available to Prime. If there is a query error, it returns an
+// error as well.
 func checkAvailabilityToPrime(event *Event) (bool, error) {
 	mtoChecker := movetaskorder.NewMoveTaskOrderChecker(event.DBConnection)
 	availableToPrime, err := mtoChecker.MTOAvailableToPrime(event.MtoID)
@@ -146,6 +72,7 @@ func checkAvailabilityToPrime(event *Event) (bool, error) {
 
 }
 
+// paymentRequestEventHandler handles all events pertaining to paymentRequests
 func paymentRequestEventHandler(event *Event) (bool, error) {
 	// CHECK SOURCE
 	// Continue only if source of event is not Prime
@@ -191,32 +118,30 @@ func paymentRequestEventHandler(event *Event) (bool, error) {
 // For alerting ALL errors should be logged here.
 func NotificationEventHandler(event *Event) error {
 
-	var logger = event.logger
-
-	// Currently it logs information about the event. Eventually it will create an entry
-	// in the notification table in the database
-	logger.Info("event.NotificationEventHandler: Running with event:",
-		zap.String("endpoint", string(event.EndpointKey)),
-		zap.String("mtoID", event.MtoID.String()),
-		zap.String("objectID", event.UpdatedObjectID.String()))
-
-	//Get the type of model which is stored in the eventType
+	//Get the model type of the updated logical object
 	modelBeingUpdated, err := GetModelFromEvent(event.EventKey)
 	if err != nil {
+		event.logger.Error("event.NotificationEventHandler: EventKey does not exist.")
 		return err
 	}
 
-	// Based on which model was updated, construct the proper payload
+	// Based on the model type, call the appropriate handler
+	var stored bool
 	switch modelBeingUpdated.(type) {
 	case models.PaymentRequest:
-		stored, err := paymentRequestEventHandler(event)
-		if err != nil {
-			event.logger.Error("event.NotificationEventHandler: ", zap.Error(err))
-			return err
-		} else if !stored {
-			event.logger.Info("event.NotificationEventHandler: No notification created.")
-		}
-		event.logger.Info("SUCCESS!")
+		stored, err = paymentRequestEventHandler(event)
+	default:
+		event.logger.Error("event.NotificationEventHandler: Unknown logical object being updated.")
+	}
+
+	// Log what happened.
+	if err != nil {
+		event.logger.Error("event.NotificationEventHandler: ", zap.Error(err))
+		return err
+	} else if !stored {
+		event.logger.Info("event.NotificationEventHandler: No notification needed to be created.")
+	} else {
+		event.logger.Info("event.NotificationEventHandler: Notification created and stored.")
 	}
 
 	return nil
